fix(fonts): convert fonts of svg to outlines
this makes sure that the svg is usable, even when the font isn't installed on one's device. Unfortunately this isn't really possible for the emoji, so that will still show as either the apple emoji or nothing, I didn't find a proper setting for it in Sketch.

I didn't commit the sketch file, since converting text to outlines makes it super hard to edit the text afterwards (there should really be an export setting in sketch to export as outlines, but alas). An option would be to have both the outlines and the text in sketch, and put the original text as a hidden layer. If that's prefered, I can do that too.

Have a nice day!